+++ payment/build.gradle
@@ -1,3 +1,7 @@
 plugins {
     id 'common-conventions'
 }
+
+dependencies {
+    implementation project(':common:response')
+}
