--- payment/src/main/java/com/atticket/PaymentApplication.java
@@ -1,11 +1,0 @@
-package com.atticket;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PaymentApplication {
-	public static void main(String[] args) {
-		SpringApplication.run(PaymentApplication.class, args);
-	}
-}
