+++ payment/src/main/java/com/atticket/payment/PaymentApplication.java
@@ -1,11 +1,11 @@
-package com.atticket;
+package com.atticket.payment;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class ProductApplication {
+public class PaymentApplication {
 	public static void main(String[] args) {
-		SpringApplication.run(ProductApplication.class, args);
+		SpringApplication.run(PaymentApplication.class, args);
 	}
 }
