+++ product/build.gradle
@@ -3,5 +3,5 @@ plugins {
 }
 
 dependencies {
-    implementation project(':common')
+    implementation project(':common:response')
 }
