+++ reservation/src/main/java/com/atticket/reservation/ReservationApplication.java
@@ -1,4 +1,4 @@
-package com.atticket;
+package com.atticket.reservation;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
