+++ settings.gradle
@@ -8,4 +8,7 @@
  */
 
 rootProject.name = 'at_ticket'
-include ':reservation', ':payment', ':common', ':product'
+include ':reservation'
+include ':payment'
+include ':product'
+include ':common:response'
